refactor: move HTTP client to pkg/client and version to pkg/version
- pkg/client: Client, ClassicClient, PlatformClient types + New() constructor
  and newRestyClient; tests moved from cmd to package-level
- pkg/version: Version var (ldflags target updated in makefile + goreleaser)
- cmd/root.go: setupClient() consolidates credential + verbosity resolution,
  mirroring dtctl's SetupClient pattern; cmd handlers call setupClient() directly
- cmd/client.go + cmd/client_test.go removed (no longer needed)
- openspec/config.yaml: project structure updated to reflect new packages
