+++ .goreleaser.yaml
@@ -18,7 +18,7 @@ builds:
       - amd64
       - arm64
     ldflags:
-      - -s -w -X github.com/dynatrace-oss/dtwiz/cmd.Version={{.Version}}
+      - -s -w -X github.com/dynatrace-oss/dtwiz/pkg/version.Version={{.Version}}
 
 archives:
   - formats:
