+++ cmd/version.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/dynatrace-oss/dtwiz/pkg/version"
 	"github.com/spf13/cobra"
 )
 
@@ -11,6 +12,6 @@ var versionCmd = &cobra.Command{
 	Short: "Print the dtwiz version",
 	Args:  cobra.NoArgs,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Printf("dtwiz %s\n", Version)
+		fmt.Printf("dtwiz %s\n", version.Version)
 	},
 }
