+++ makefile
@@ -6,7 +6,7 @@ VERSION := $(shell git describe --tags --always --dirty 2>/dev/null || echo dev)
 MD_LINT_CLI_IMAGE := "ghcr.io/igorshubovych/markdownlint-cli:v0.31.1"
 
 build:
-	$(GO) build -ldflags "-X github.com/dynatrace-oss/dtwiz/cmd.Version=$(VERSION)" -o $(BINARY) .
+	$(GO) build -ldflags "-X github.com/dynatrace-oss/dtwiz/pkg/version.Version=$(VERSION)" -o $(BINARY) .
 
 install:
 	$(GO) install .
