+++ pkg/client/client_test.go
@@ -1,4 +1,4 @@
-package cmd
+package client
 
 import (
 	"net/http"
@@ -6,6 +6,8 @@ import (
 	"sync/atomic"
 	"testing"
 	"time"
+
+	"github.com/dynatrace-oss/dtwiz/pkg/version"
 )
 
 func TestSensitiveHTTPHeaders(t *testing.T) {
@@ -34,8 +36,8 @@ func TestNewRestyClient_Settings(t *testing.T) {
 	if got := rc.Header.Get("Authorization"); got != "Api-Token tok" {
 		t.Errorf("Authorization = %q, want %q", got, "Api-Token tok")
 	}
-	if got := rc.Header.Get("User-Agent"); got != "dtwiz/"+Version {
-		t.Errorf("User-Agent = %q, want %q", got, "dtwiz/"+Version)
+	if got := rc.Header.Get("User-Agent"); got != "dtwiz/"+version.Version {
+		t.Errorf("User-Agent = %q, want %q", got, "dtwiz/"+version.Version)
 	}
 	if got := rc.Header.Get("Accept-Encoding"); got != "gzip" {
 		t.Errorf("Accept-Encoding = %q, want gzip", got)
@@ -104,62 +106,5 @@ func TestNewRestyClient_NoRetryOn4xx(t *testing.T) {
 	}
 	if n := calls.Load(); n != 1 {
 		t.Errorf("server calls = %d, want 1 (no retry on 404)", n)
-	}
-}
-
-func TestNewHTTPClient_MissingEnvironment(t *testing.T) {
-	t.Setenv("DT_ENVIRONMENT", "")
-	t.Setenv("DT_ACCESS_TOKEN", "")
-	t.Setenv("DT_PLATFORM_TOKEN", "")
-	environmentFlag, accessTokenFlag, platformTokenFlag = "", "", ""
-
-	if _, err := NewHTTPClient(); err == nil {
-		t.Fatal("expected error when DT_ENVIRONMENT is not set")
-	}
-}
-
-func TestNewHTTPClient_MissingAccessToken(t *testing.T) {
-	t.Setenv("DT_ENVIRONMENT", "https://abc123.apps.dynatrace.com")
-	t.Setenv("DT_ACCESS_TOKEN", "")
-	t.Setenv("DT_PLATFORM_TOKEN", "")
-	environmentFlag, accessTokenFlag, platformTokenFlag = "", "", ""
-
-	if _, err := NewHTTPClient(); err == nil {
-		t.Fatal("expected error when DT_ACCESS_TOKEN is not set")
-	}
-}
-
-func TestNewHTTPClient_MissingPlatformToken(t *testing.T) {
-	t.Setenv("DT_ENVIRONMENT", "https://abc123.apps.dynatrace.com")
-	t.Setenv("DT_ACCESS_TOKEN", "dt0c01.test")
-	t.Setenv("DT_PLATFORM_TOKEN", "")
-	environmentFlag, accessTokenFlag, platformTokenFlag = "", "", ""
-
-	if _, err := NewHTTPClient(); err == nil {
-		t.Fatal("expected error when DT_PLATFORM_TOKEN is not set")
-	}
-}
-
-func TestNewHTTPClient_Success(t *testing.T) {
-	t.Setenv("DT_ENVIRONMENT", "https://abc123.apps.dynatrace.com")
-	t.Setenv("DT_ACCESS_TOKEN", "dt0c01.test")
-	t.Setenv("DT_PLATFORM_TOKEN", "dt0s16.test")
-	environmentFlag, accessTokenFlag, platformTokenFlag = "", "", ""
-
-	c, err := NewHTTPClient()
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if c.Classic == nil || c.Platform == nil {
-		t.Fatal("Classic or Platform is nil")
-	}
-
-	const wantClassic = "https://abc123.live.dynatrace.com"
-	if got := c.Classic.BaseURL(); got != wantClassic {
-		t.Errorf("Classic.BaseURL() = %q, want %q", got, wantClassic)
-	}
-	const wantPlatform = "https://abc123.apps.dynatrace.com"
-	if got := c.Platform.BaseURL(); got != wantPlatform {
-		t.Errorf("Platform.BaseURL() = %q, want %q", got, wantPlatform)
 	}
 }
