+++ pkg/version/version.go
@@ -1,0 +1,4 @@
+package version
+
+// Version is set at build time via -ldflags.
+var Version = "dev"
